+++ client/app/components/userData/FavoriteButton.tsx
@@ -1,8 +1,8 @@
 "use client";
-
-import { useState, useTransition } from "react";
+import { useState } from "react";
 import { Heart } from "lucide-react";
-import { addFavoriteSchool } from "@/mypage/actions";
+import { useRouter } from "next/navigation";
+import { addFavoriteSchool } from "@/actions/userPage";
 
 export default function FavoriteButton({
   schoolId,
@@ -14,29 +14,48 @@ export default function FavoriteButton({
   className?: string;
 }) {
   const [isFavorite, setIsFavorite] = useState(isInitiallyFavorite);
-  const [isPending, startTransition] = useTransition();
+  const [isPending, setIsPending] = useState(false);
+  const router = useRouter();
 
-  const handleClick = () => {
-    if (isFavorite) return;
-    startTransition(async () => {
-      await addFavoriteSchool(schoolId);
+  // フォーム送信ハンドラ（Server Actionを呼び出す）
+  const handleAddFavorite = async (formData: FormData) => {
+    setIsPending(true);
+    try {
+      // Server Actionを呼び出す
+      await addFavoriteSchool(formData.get("schoolId") as string);
+      // 状態を更新
       setIsFavorite(true);
-    });
+      // UIを更新（必要に応じて）
+      router.refresh();
+    } catch (error) {
+      console.error("お気に入り追加エラー:", error);
+    } finally {
+      setIsPending(false);
+    }
   };
 
+  // すでにお気に入り済みか登録中の場合はボタンを無効化
+  const isDisabled = isFavorite || isPending;
+
   return (
-    <button
-      type="button"
-      onClick={handleClick}
-      disabled={isFavorite || isPending}
-      className={`py-3 px-6 bg-white border border-gray-300 text-sm text-gray-700 rounded-md flex items-center justify-center hover:bg-gray-50 disabled:hover:bg-white disabled:opacity-60 ${className}`}
-    >
-      <Heart className="w-4 h-4 mr-2 text-gray-600" />
-      {isFavorite
-        ? "気になる済み"
-        : isPending
-        ? "登録中..."
-        : "気になるに登録する"}
-    </button>
+    <form action={handleAddFavorite}>
+      <input type="hidden" name="schoolId" value={schoolId} />
+      <button
+        type="submit"
+        disabled={isDisabled}
+        className={`py-3 px-6 bg-white border border-gray-300 text-sm text-gray-700 rounded-md flex items-center justify-center hover:bg-gray-50 disabled:hover:bg-white disabled:opacity-60 ${className}`}
+      >
+        <Heart
+          className={`w-4 h-4 mr-2 ${
+            isFavorite ? "text-gray-500 fill-gray-500" : "text-gray-600"
+          }`}
+        />
+        {isFavorite
+          ? "気になる済み"
+          : isPending
+          ? "登録中..."
+          : "気になるに登録する"}
+      </button>
+    </form>
   );
 }
